Correction to sync endpoint
