+++ app/services/legacy/legacy_to_beta_sync_service.rb
@@ -1,60 +1,48 @@
-class Legacy::LegacyToBetaSyncService
-  attr_reader :dataset
+require 'util/metadata_tools'
 
-  def initialize(dataset)
-    @dataset = dataset
+class Legacy::BetaToLegacySyncService
+  ENDPOINTS = {
+    modified_datasets: 'api/3/action/package_search?q=metadata_modified:[NOW-1DAY%20TO%20NOW]&rows=5000'.freeze,
+    new_datasets: 'api/3/action/package_search?q=metadata_created:[NOW-1DAY%20TO%20NOW]&rows=5000'.freeze
+  }
+
+  def initialize(orgs_cache:, theme_cache:, logger:, legacy_server:)
+    @orgs_cache = orgs_cache
+    @theme_cache = theme_cache
+    @logger = logger
+    @legacy_server = legacy_server
+    @count = 0
   end
 
-  def sync
-    sync_dataset
-    sync_datafiles
+  def run
+    @logger.info "Importing legacy datasets...\r"
+    [modified_datasets, new_datasets].each do |datasets|
+      datasets.fetch('result', {}).fetch('results', []).each do |dataset|
+        import dataset
+      end
+    end
+    @logger.info "Imported #{@count} datasets...\r"
   end
 
   private
 
-  def sync_dataset
-    if dataset.published_date.present?
-      update_legacy_dataset
-    else
-      create_legacy_dataset
+  def modified_datasets
+    @legacy_server.get ENDPOINTS[:modified_datasets]
+  end
+
+  def new_datasets
+    @legacy_server.get ENDPOINTS[:new_datasets]
+  end
+
+  def import(dataset)
+    begin
+      @logger.info "Attempting to save legacy dataset to postgres and elasticsearch - legacy_id: #{dataset['id']}"
+      MetadataTools.persist(dataset, @orgs_cache, @theme_cache)
+      MetadataTools.index(dataset)
+      @logger.info "Legacy dataset saved - legacy_id: #{dataset['id']}"
+    rescue => e
+      Raven.capture_exception e.message
     end
-  end
-
-  def sync_datafiles
-    create_legacy_datafiles
-    update_legacy_datafiles
-  end
-
-  def update_legacy_dataset
-    ::LegacyDatasetUpdateWorker.perform_async(dataset.id)
-  end
-
-  def create_legacy_dataset
-    ::LegacyDatasetCreateWorker.perform_async(dataset.id)
-  end
-
-  def create_legacy_datafiles
-    datafiles_created_since_latest_publication.each do |datafile|
-      ::LegacyDatafileCreateWorker.perform_async(datafile.id)
-    end
-  end
-
-  def update_legacy_datafiles
-    datafiles_modified_since_latest_publication.each do |datafile|
-      ::LegacyDatafileUpdateWorker.perform_async(datafile.id)
-    end
-  end
-
-  def datafiles_created_since_latest_publication
-    Datafile
-      .where(dataset_id: dataset.id)
-      .created_after_date(dataset.last_published_at)
-  end
-
-  def datafiles_modified_since_latest_publication
-    Datafile
-      .where(dataset_id: dataset.id)
-      .created_before_date(dataset.last_published_at)
-      .updated_after_date(dataset.last_published_at)
+    @count += 1
   end
 end
